+++ lib/hooks/use-front-chat-boot/index.ts
@@ -1,7 +1,7 @@
 import {useEffect, useRef, useState} from 'react';
 
 import {boot} from '../../helpers/boot';
-import {type FrontChat, type FrontChatOptions, type FrontChatParams} from '../../types';
+import {type FrontChat, type FrontChatOptions, type FrontChatParams} from '../../types/types';
 
 /*
  * Types.
