+++ lib/index.ts
@@ -1,5 +1,5 @@
-import * as helpers from './helpers/index';
-import * as hooks from './hooks/index';
-import * as types from './types/index';
+import * as helpers from './helpers';
+import * as hooks from './hooks';
+import * as types from './types';
 
 export {helpers, hooks, types};
