+++ lib/types/types.ts
@@ -1,0 +1,23 @@
+/*
+ * Global types.
+ */
+
+declare global {
+  interface Window {
+    FrontChat: FrontChat | undefined;
+  }
+}
+
+/*
+ * Types.
+ */
+
+export interface FrontChatOptions {
+  nonce?: string;
+}
+
+type UnbindHandler = () => void;
+
+// TODO: Fix this type to match front-client.
+export type FrontChatParams = Record<string, string | boolean | object>;
+export type FrontChat = (cmdType: string, params?: FrontChatParams) => UnbindHandler | undefined;
