+++ tsup.config.ts
@@ -4,13 +4,16 @@ export default defineConfig({
   entry: {
     index: 'lib/index.ts',
     'helpers/index': 'lib/helpers/index.ts',
+    // 'helpers/boot/index': 'lib/helpers/boot/index.ts',
     'hooks/index': 'lib/hooks/index.ts',
+    // 'hooks/use-front-chat/index': 'lib/hooks/use-front-chat/index.ts',
+    // 'hooks/use-front-chat-boot/index': 'lib/hooks/use-front-chat-boot/index.ts',
     'types/index': 'lib/types/index.ts'
   },
   outDir: 'dist',
-  format: ['esm', 'cjs'],
+  format: ['cjs', 'esm'],
   clean: true,
   sourcemap: true,
-  splitting: false,
+  treeshake: true,
   dts: true
 });
